+++ node/typings.d.ts
@@ -43,6 +43,7 @@ interface OrganizationInput {
   tradeName?: string
   b2bCustomerAdmin: B2BCustomerInput
   defaultCostCenter: DefaultCostCenterInput
+  customFields?: CustomField[]
 }
 
 interface B2BCustomerInput {
@@ -56,6 +57,7 @@ interface DefaultCostCenterInput {
   address: AddressInput
   phoneNumber?: string
   businessDocument?: string
+  customFields?: CustomField[]
   stateRegistration?: string
 }
 
@@ -65,6 +67,7 @@ interface CostCenterInput {
   paymentTerms?: PaymentTerm[]
   phoneNumber?: string
   businessDocument?: string
+  customFields?: CustomField[]
   stateRegistration?: string
 }
 
@@ -100,6 +103,7 @@ interface Organization {
   paymentTerms: PaymentTerm[]
   status: string
   created: string
+  customFields?: CustomField[]
 }
 
 interface CostCenter {
@@ -110,6 +114,7 @@ interface CostCenter {
   paymentTerms: PaymentTerm[]
   phoneNumber?: string
   businessDocument?: string
+  customFields: CustomField[]
 }
 
 interface Address {
@@ -130,6 +135,14 @@ interface Address {
   checked?: boolean
 }
 
+interface B2BSettings {
+  autoApprove: boolean
+  defaultPaymentTerms: PaymentTerm[]
+  defaultPriceTables: [string]
+  organizationCustomFields: SettingsCustomField[]
+  costCenterCustomFields: SettingsCustomField[]
+}
+
 interface UserArgs {
   id?: string
   roleId: string
@@ -145,4 +158,37 @@ interface UserArgs {
 interface PaymentTerm {
   name: string
   id: string
+}
+
+interface Price {
+  name: string
+  id: string
+}
+
+interface CustomField {
+  name: string
+  type: 'text' | 'dropdown'
+  value: string
+  dropdownValues?: Array<{ label: string; value: string }> | null
+  useOnRegistration?: boolean
+}
+
+interface SettingsCustomField {
+  name: string
+  type: 'text' | 'dropdown'
+  dropdownValues?: Array<{ label: string; value: string }> | null
+  useOnRegistration?: boolean
+}
+
+interface CustomFieldSetting {
+  name: string
+  type: 'text' | 'dropdown'
+}
+
+interface B2BSettingsInput {
+  autoApprove: boolean
+  defaultPaymentTerms: PaymentTerm[]
+  defaultPriceTables: Price[]
+  organizationCustomFields: CustomFieldSetting[]
+  costCenterCustomFields: CustomFieldSetting[]
 }